Initial commit
Current strategy: brute force sequential with cutoff determined by 40th
percentile of top 2 average similarities per phoneme (non-matching) over
place-dictionary word combinations. This is too slow, and by cutting off,
we get the same terms over and over again for the same leading phoneme.

A small perturbation of this technique is to randomize the cutoff, so we
get slightly different instances. However, it's monotonic in the fidelity
of the final result, so besides adding some spice to the result, it isn't
effective.